move bootloader runner to a nicer `lib.rs`
