+++ platforms/x86_64/bootloader/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+anyhow = "1"
 # used for UEFI booting in QEMU
 ovmf-prebuilt = "0.1.0-alpha.1"
 # used for the QEMU runner
